extract dynamo helpers
