+++ tt-media-server/cpp_server/tests/e2e/prefix_cache_e2e_test.cpp
@@ -15,274 +15,20 @@
 // Usage:
 //   ./prefix_cache_e2e_test
 
-#include <arpa/inet.h>
 #include <gtest/gtest.h>
-#include <json/json.h>
-#include <netdb.h>
-#include <sys/socket.h>
-#include <unistd.h>
 
-#include <chrono>
 #include <cmath>
-#include <cstdint>
-#include <cstdlib>
-#include <cstring>
-#include <fstream>
 #include <iostream>
-#include <sstream>
-#include <stdexcept>
 #include <string>
 #include <thread>
 #include <vector>
 
 #include "config/defaults.hpp"
+#include "dynamo_test_helpers.hpp"
 
 namespace {
 
-// ---------------------------------------------------------------------------
-// Docker / container helpers
-// ---------------------------------------------------------------------------
-
-std::string detectDockerGateway() {
-  std::ifstream route("/proc/net/route");
-  if (!route) return "127.0.0.1";
-
-  std::string line;
-  std::getline(route, line);  // skip header
-  while (std::getline(route, line)) {
-    std::istringstream iss(line);
-    std::string iface, dest, gateway;
-    if (iss >> iface >> dest >> gateway && dest == "00000000") {
-      unsigned int gw = std::stoul(gateway, nullptr, 16);
-      unsigned char* bytes = reinterpret_cast<unsigned char*>(&gw);
-      return std::to_string(bytes[0]) + "." + std::to_string(bytes[1]) + "." +
-             std::to_string(bytes[2]) + "." + std::to_string(bytes[3]);
-    }
-  }
-  return "127.0.0.1";
-}
-
-// ---------------------------------------------------------------------------
-// SSE response parsing
-// ---------------------------------------------------------------------------
-
-struct UsageInfo {
-  int promptTokens = 0;
-  int completionTokens = 0;
-  int totalTokens = 0;
-  int cachedTokens = 0;
-};
-
-struct ChatResponse {
-  int statusCode = 0;
-  std::string content;
-  UsageInfo usage;
-  std::string error;
-  bool ok() const { return statusCode == 200 && error.empty(); }
-};
-
-int parseStatusCode(const std::string& response) {
-  auto pos = response.find(' ');
-  if (pos == std::string::npos) return 0;
-  return std::stoi(response.substr(pos + 1, 3));
-}
-
-ChatResponse parseStreamingResponse(const std::string& rawResponse) {
-  ChatResponse result;
-  result.statusCode = parseStatusCode(rawResponse);
-
-  if (result.statusCode != 200) {
-    result.error = "HTTP " + std::to_string(result.statusCode);
-    return result;
-  }
-
-  std::istringstream stream(rawResponse);
-  std::string line;
-
-  while (std::getline(stream, line)) {
-    if (!line.empty() && line.back() == '\r') {
-      line.pop_back();
-    }
-
-    if (line.rfind("data: ", 0) != 0) continue;
-
-    std::string data = line.substr(6);
-    if (data == "[DONE]") break;
-
-    Json::Value chunk;
-    Json::CharReaderBuilder builder;
-    std::unique_ptr<Json::CharReader> reader(builder.newCharReader());
-    std::string errors;
-
-    if (!reader->parse(data.c_str(), data.c_str() + data.size(), &chunk,
-                       &errors)) {
-      continue;
-    }
-
-    if (chunk.isMember("choices") && chunk["choices"].isArray() &&
-        !chunk["choices"].empty()) {
-      const auto& delta = chunk["choices"][0]["delta"];
-      if (delta.isMember("content") && !delta["content"].isNull()) {
-        result.content += delta["content"].asString();
-      }
-      if (delta.isMember("reasoning_content") &&
-          !delta["reasoning_content"].isNull()) {
-        result.content += delta["reasoning_content"].asString();
-      }
-    }
-
-    if (chunk.isMember("usage") && chunk["usage"].isObject()) {
-      const auto& usage = chunk["usage"];
-      result.usage.promptTokens = usage.get("prompt_tokens", 0).asInt();
-      result.usage.completionTokens = usage.get("completion_tokens", 0).asInt();
-      result.usage.totalTokens = usage.get("total_tokens", 0).asInt();
-
-      if (usage.isMember("prompt_tokens_details")) {
-        const auto& ptd = usage["prompt_tokens_details"];
-        result.usage.cachedTokens = ptd.get("cached_tokens", 0).asInt();
-      }
-    }
-  }
-
-  return result;
-}
-
-// ---------------------------------------------------------------------------
-// HTTP client
-// ---------------------------------------------------------------------------
-
-std::string buildHttpRequest(const std::string& host, uint16_t port,
-                             const std::string& body) {
-  std::ostringstream oss;
-  oss << "POST /v1/chat/completions HTTP/1.1\r\n"
-      << "Host: " << host << ":" << port << "\r\n"
-      << "Content-Type: application/json\r\n"
-      << "Content-Length: " << body.size() << "\r\n"
-      << "\r\n"
-      << body;
-  return oss.str();
-}
-
-std::string sendHttpRequest(const std::string& host, uint16_t port,
-                            const std::string& body, int timeoutMs = 120000) {
-  int sock = ::socket(AF_INET, SOCK_STREAM, 0);
-  if (sock < 0) {
-    throw std::runtime_error("Failed to create socket");
-  }
-
-  sockaddr_in addr{};
-  addr.sin_family = AF_INET;
-  addr.sin_port = htons(port);
-
-  if (::inet_pton(AF_INET, host.c_str(), &addr.sin_addr) <= 0) {
-    struct hostent* he = ::gethostbyname(host.c_str());
-    if (!he) {
-      ::close(sock);
-      throw std::runtime_error("Failed to resolve host: " + host);
-    }
-    std::memcpy(&addr.sin_addr, he->h_addr_list[0], he->h_length);
-  }
-
-  if (::connect(sock, reinterpret_cast<sockaddr*>(&addr), sizeof(addr)) != 0) {
-    ::close(sock);
-    throw std::runtime_error("Failed to connect to " + host + ":" +
-                             std::to_string(port));
-  }
-
-  std::string request = buildHttpRequest(host, port, body);
-  if (::send(sock, request.c_str(), request.size(), 0) < 0) {
-    ::close(sock);
-    throw std::runtime_error("Failed to send request");
-  }
-
-  timeval tv{timeoutMs / 1000, (timeoutMs % 1000) * 1000};
-  ::setsockopt(sock, SOL_SOCKET, SO_RCVTIMEO, &tv, sizeof(tv));
-
-  std::string response;
-  char buf[4096];
-  ssize_t n;
-
-  while ((n = ::recv(sock, buf, sizeof(buf), 0)) > 0) {
-    response.append(buf, static_cast<size_t>(n));
-    if (response.find("data: [DONE]") != std::string::npos) {
-      break;
-    }
-  }
-
-  ::close(sock);
-  return response;
-}
-
-// ---------------------------------------------------------------------------
-// JSON request building
-// ---------------------------------------------------------------------------
-
-std::string buildChatRequestJson(const std::string& model,
-                                 const std::vector<Json::Value>& messages,
-                                 int maxTokens = 32, bool stream = true) {
-  Json::Value root;
-  root["model"] = model;
-  root["max_tokens"] = maxTokens;
-  root["stream"] = stream;
-
-  if (stream) {
-    Json::Value streamOptions;
-    streamOptions["include_usage"] = true;
-    root["stream_options"] = streamOptions;
-  }
-
-  Json::Value messagesArray(Json::arrayValue);
-  for (const auto& msg : messages) {
-    messagesArray.append(msg);
-  }
-  root["messages"] = messagesArray;
-
-  Json::StreamWriterBuilder writer;
-  writer["indentation"] = "";
-  return Json::writeString(writer, root);
-}
-
-Json::Value makeMessage(const std::string& role, const std::string& content) {
-  Json::Value msg;
-  msg["role"] = role;
-  msg["content"] = content;
-  return msg;
-}
-
-// ---------------------------------------------------------------------------
-// Server readiness helpers
-// ---------------------------------------------------------------------------
-
-bool waitForTcpPort(const std::string& host, uint16_t port,
-                    int timeoutSec = 60) {
-  auto deadline =
-      std::chrono::steady_clock::now() + std::chrono::seconds(timeoutSec);
-
-  while (std::chrono::steady_clock::now() < deadline) {
-    try {
-      int sock = ::socket(AF_INET, SOCK_STREAM, 0);
-      if (sock < 0) continue;
-
-      sockaddr_in addr{};
-      addr.sin_family = AF_INET;
-      addr.sin_port = htons(port);
-      ::inet_pton(AF_INET, host.c_str(), &addr.sin_addr);
-
-      timeval tv{2, 0};
-      ::setsockopt(sock, SOL_SOCKET, SO_RCVTIMEO, &tv, sizeof(tv));
-      ::setsockopt(sock, SOL_SOCKET, SO_SNDTIMEO, &tv, sizeof(tv));
-
-      bool connected = ::connect(sock, reinterpret_cast<sockaddr*>(&addr),
-                                 sizeof(addr)) == 0;
-      ::close(sock);
-
-      if (connected) return true;
-    } catch (...) {
-    }
-    std::this_thread::sleep_for(std::chrono::milliseconds(500));
-  }
-  return false;
-}
+using namespace tt::test::dynamo;
 
 // ---------------------------------------------------------------------------
 // Prefix cache calculation
@@ -302,21 +48,19 @@ int computeExpectedCachedTokens(int promptTokens, size_t firstBlockSize,
 }
 
 // ---------------------------------------------------------------------------
-// Test configuration
+// Test configuration (extends DynamoConfig with KV cache settings)
 // ---------------------------------------------------------------------------
 
-struct TestConfig {
-  std::string host = detectDockerGateway();
-  uint16_t port = 8080;
-  std::string model = "deepseek-ai/DeepSeek-R1-0528";
+struct PrefixCacheTestConfig {
+  DynamoConfig dynamo;
   size_t firstBlockSize = tt::config::defaults::KV_CACHE_FIRST_BLOCK_SIZE;
   size_t blockSize = tt::config::defaults::KV_CACHE_BLOCK_SIZE;
 
-  static TestConfig fromEnv() {
-    TestConfig cfg;
-    if (const char* h = std::getenv("DYNAMO_HOST")) cfg.host = h;
-    if (const char* p = std::getenv("DYNAMO_PORT")) cfg.port = std::stoi(p);
-    if (const char* m = std::getenv("DYNAMO_MODEL")) cfg.model = m;
+  static PrefixCacheTestConfig fromEnv() {
+    PrefixCacheTestConfig cfg;
+    cfg.dynamo = DynamoConfig::fromEnv();
+    cfg.dynamo.model = "deepseek-ai/DeepSeek-R1-0528";
+    if (const char* m = std::getenv("DYNAMO_MODEL")) cfg.dynamo.model = m;
     if (const char* fb = std::getenv("KV_CACHE_FIRST_BLOCK_SIZE"))
       cfg.firstBlockSize = std::stoul(fb);
     if (const char* bs = std::getenv("KV_CACHE_BLOCK_SIZE"))
@@ -325,44 +69,13 @@ struct TestConfig {
   }
 };
 
-ChatResponse sendChat(const TestConfig& cfg,
-                      const std::vector<Json::Value>& messages,
-                      int maxTokens = 32) {
-  std::string body = buildChatRequestJson(cfg.model, messages, maxTokens);
-  try {
-    std::string rawResponse = sendHttpRequest(cfg.host, cfg.port, body);
-    return parseStreamingResponse(rawResponse);
-  } catch (const std::exception& e) {
-    ChatResponse result;
-    result.error = e.what();
-    return result;
-  }
-}
-
-bool warmupServer(const TestConfig& cfg) {
-  std::vector<Json::Value> warmupMessages = {
-      makeMessage("system", "You are a helpful assistant."),
-      makeMessage("user", "Say hello.")};
-  for (int attempt = 0; attempt < 5; ++attempt) {
-    ChatResponse r = sendChat(cfg, warmupMessages, 8);
-    if (r.ok()) {
-      std::cout << "Warmup succeeded after " << (attempt + 1) << " attempt(s)"
-                << std::endl;
-      return true;
-    }
-    std::cout << "Warmup attempt " << (attempt + 1) << " failed: " << r.error
-              << std::endl;
-    std::this_thread::sleep_for(std::chrono::seconds(2));
-  }
-  return false;
-}
-
+// ---------------------------------------------------------------------------
 // System prompts that fill at least one hash block (~200+ tokens).
 // Each prompt starts with completely different text to ensure no prefix overlap
 // between tests. The unique suffix added at runtime ensures no overlap with
 // prior test runs.
+// ---------------------------------------------------------------------------
 
-// Test 1: Coding assistant theme
 const char* kSystemPromptCoding =
     "CODING ASSISTANT ALPHA: You are a highly capable AI coding assistant "
     "working inside an IDE. You have access to the full project source tree "
@@ -383,7 +96,6 @@ const char* kSystemPromptCoding =
     "RESTful conventions and provide clear error messages with appropriate "
     "HTTP status codes.";
 
-// Test 2: Marine biology theme (completely different prefix)
 const char* kSystemPromptMarine =
     "MARINE BIOLOGY EXPERT: You specialize in oceanography and marine life. "
     "Your expertise covers coral reef ecosystems, deep sea creatures, whale "
@@ -403,7 +115,6 @@ const char* kSystemPromptMarine =
     "Consider the economic importance of fisheries and sustainable harvesting "
     "practices.";
 
-// Test 3: Astronomy theme (completely different prefix)
 const char* kSystemPromptAstronomy =
     "ASTRONOMY SPECIALIST: You are an expert in astrophysics and space "
     "exploration. Your knowledge spans stellar evolution, galaxy formation, "
@@ -426,56 +137,60 @@ const char* kSystemPromptAstronomy =
 class PrefixCacheE2ETest : public ::testing::Test {
  protected:
   static void SetUpTestSuite() {
-    cfg = TestConfig::fromEnv();
-    std::cout << "Prefix cache E2E test against " << cfg.host << ":" << cfg.port
-              << std::endl;
-    std::cout << "  model=" << cfg.model
-              << "  firstBlockSize=" << cfg.firstBlockSize
-              << "  blockSize=" << cfg.blockSize << std::endl;
+    cfg_ = PrefixCacheTestConfig::fromEnv();
+    client_ = std::make_unique<DynamoClient>(cfg_.dynamo);
+
+    std::cout << "Prefix cache E2E test against " << cfg_.dynamo.host << ":"
+              << cfg_.dynamo.port << std::endl;
+    std::cout << "  model=" << cfg_.dynamo.model
+              << "  firstBlockSize=" << cfg_.firstBlockSize
+              << "  blockSize=" << cfg_.blockSize << std::endl;
     std::cout << "Waiting for server..." << std::endl;
 
-    ASSERT_TRUE(waitForTcpPort(cfg.host, cfg.port))
-        << "Server not ready within timeout";
+    ASSERT_TRUE(client_->waitForServer()) << "Server not ready within timeout";
     std::cout << "Server ready, warming up..." << std::endl;
 
-    ASSERT_TRUE(warmupServer(cfg))
+    ASSERT_TRUE(client_->warmup())
         << "Server warmup failed (Dynamo frontend may not have discovered "
            "backends)";
     std::cout << "Server warmed up." << std::endl;
   }
 
-  static TestConfig cfg;
+  static void TearDownTestSuite() { client_.reset(); }
+
+  ChatResponse sendChat(const std::vector<Json::Value>& messages,
+                        int maxTokens = 32) {
+    return client_->sendChat(messages, maxTokens);
+  }
+
+  static PrefixCacheTestConfig cfg_;
+  static std::unique_ptr<DynamoClient> client_;
 };
 
-TestConfig PrefixCacheE2ETest::cfg;
+PrefixCacheTestConfig PrefixCacheE2ETest::cfg_;
+std::unique_ptr<DynamoClient> PrefixCacheE2ETest::client_;
 
 TEST_F(PrefixCacheE2ETest, CacheReplayScenario) {
   // Test cache behavior across multiple requests including replays.
   //
-  // 1. Request 1: Fresh prompt → cached_tokens = 0
-  // 2. Request 2: Continuation → cached_tokens grows
-  // 3. Request 3: Different prompt → cached_tokens = 0
-  // 4. Request 4: Replay R1 → cached_tokens = block-aligned(R1 prompt)
-  // 5. Request 5: Replay R2 → cached_tokens = block-aligned(R2 prompt)
+  // 1. Request 1: Fresh prompt -> cached_tokens = 0
+  // 2. Request 2: Continuation -> cached_tokens grows
+  // 3. Request 3: Different prompt -> cached_tokens = 0
+  // 4. Request 4: Replay R1 -> cached_tokens = block-aligned(R1 prompt)
+  // 5. Request 5: Replay R2 -> cached_tokens = block-aligned(R2 prompt)
 
   std::cout << "\n=== Test: Cache replay scenario ===" << std::endl;
 
-  // Unique suffix ensures no overlap with prior test runs
-  auto now = std::chrono::system_clock::now();
-  auto epoch = now.time_since_epoch();
-  auto millis =
-      std::chrono::duration_cast<std::chrono::milliseconds>(epoch).count();
-  std::string uniqueSuffix = " [replay-test-" + std::to_string(millis) + "]";
+  std::string uniqueSuffix =
+      " [" + generateUniqueTestId("replay-test") + "]";
 
-  // -------------------------------------------------------------------------
   // Request 1: Fresh prompt (coding assistant theme)
-  // -------------------------------------------------------------------------
   std::vector<Json::Value> r1Messages = {
       makeMessage("system", std::string(kSystemPromptCoding) + uniqueSuffix),
       makeMessage("user", "What is the capital of France?")};
 
   std::cout << "  Request 1 (fresh prompt)..." << std::endl;
-  ChatResponse r1 = sendChat(cfg, r1Messages);
+  ChatResponse r1 = sendChat(r1Messages);
   ASSERT_TRUE(r1.ok()) << "Request 1 failed: " << r1.error;
   std::cout << "    prompt=" << r1.usage.promptTokens
             << " cached=" << r1.usage.cachedTokens
@@ -486,9 +201,7 @@ TEST_F(PrefixCacheE2ETest, CacheReplayScenario) {
 
   std::vector<Json::Value> r1MessagesCopy = r1Messages;
 
-  // -------------------------------------------------------------------------
   // Request 2: Continuation of R1
-  // -------------------------------------------------------------------------
   std::vector<Json::Value> r2Messages = r1Messages;
   r2Messages.push_back(makeMessage("assistant", r1.content));
   r2Messages.push_back(
@@ -497,18 +210,15 @@ TEST_F(PrefixCacheE2ETest, CacheReplayScenario) {
   std::this_thread::sleep_for(std::chrono::milliseconds(300));
 
   std::cout << "  Request 2 (continuation of R1)..." << std::endl;
-  ChatResponse r2 = sendChat(cfg, r2Messages);
+  ChatResponse r2 = sendChat(r2Messages);
   ASSERT_TRUE(r2.ok()) << "Request 2 failed: " << r2.error;
   std::cout << "    prompt=" << r2.usage.promptTokens
             << " cached=" << r2.usage.cachedTokens
             << " completion=" << r2.usage.completionTokens << std::endl;
 
-  // R2's prompt includes the assistant response as text. With mock_pipeline,
-  // the tokenized assistant text round-trips to the same tokens as R1's
-  // completion, so R2 matches R1's full session (prompt + completion).
   int r1SessionTokens = r1.usage.promptTokens + r1.usage.completionTokens;
   int r2ExpectedCached = computeExpectedCachedTokens(
-      r1SessionTokens, cfg.firstBlockSize, cfg.blockSize);
+      r1SessionTokens, cfg_.firstBlockSize, cfg_.blockSize);
   std::cout << "    Expected cached: " << r2ExpectedCached << std::endl;
 
   EXPECT_GT(r2.usage.cachedTokens, 0)
@@ -516,12 +226,9 @@ TEST_F(PrefixCacheE2ETest, CacheReplayScenario) {
   EXPECT_LE(std::abs(r2.usage.cachedTokens - r2ExpectedCached), 1)
       << "Request 2 cached should match block-aligned R1 session";
 
-  // Save R2 messages for replay later
   std::vector<Json::Value> r2MessagesCopy = r2Messages;
 
-  // -------------------------------------------------------------------------
   // Request 3: Different prompt (marine biology theme, no prefix overlap)
-  // -------------------------------------------------------------------------
   std::vector<Json::Value> r3Messages = {
       makeMessage("system", std::string(kSystemPromptMarine) + uniqueSuffix),
       makeMessage("user", "Tell me about coral reef ecosystems.")};
@@ -529,7 +236,7 @@ TEST_F(PrefixCacheE2ETest, CacheReplayScenario) {
   std::this_thread::sleep_for(std::chrono::milliseconds(300));
 
   std::cout << "  Request 3 (different prompt)..." << std::endl;
-  ChatResponse r3 = sendChat(cfg, r3Messages);
+  ChatResponse r3 = sendChat(r3Messages);
   ASSERT_TRUE(r3.ok()) << "Request 3 failed: " << r3.error;
   std::cout << "    prompt=" << r3.usage.promptTokens
             << " cached=" << r3.usage.cachedTokens
@@ -538,23 +245,18 @@ TEST_F(PrefixCacheE2ETest, CacheReplayScenario) {
   EXPECT_EQ(r3.usage.cachedTokens, 0)
       << "Request 3 should have cached_tokens=0 (completely different prompt)";
 
-  // -------------------------------------------------------------------------
-  // Request 4: Replay exact R1 prompt → should hit full cache
-  // -------------------------------------------------------------------------
+  // Request 4: Replay exact R1 prompt -> should hit full cache
   std::this_thread::sleep_for(std::chrono::milliseconds(300));
 
   std::cout << "  Request 4 (replay R1 prompt)..." << std::endl;
-  ChatResponse r4 = sendChat(cfg, r1MessagesCopy);
+  ChatResponse r4 = sendChat(r1MessagesCopy);
   ASSERT_TRUE(r4.ok()) << "Request 4 failed: " << r4.error;
   std::cout << "    prompt=" << r4.usage.promptTokens
             << " cached=" << r4.usage.cachedTokens
             << " completion=" << r4.usage.completionTokens << std::endl;
 
-  // R4 replays R1's original prompt exactly. R1's session has completion tokens
-  // cached, but R4 only sends the original prompt tokens, so the match is
-  // limited to min(r4_prompt, r1_session) = r4_prompt tokens.
   int r4ExpectedCached = computeExpectedCachedTokens(
-      r4.usage.promptTokens, cfg.firstBlockSize, cfg.blockSize);
+      r4.usage.promptTokens, cfg_.firstBlockSize, cfg_.blockSize);
   std::cout << "    Expected cached: " << r4ExpectedCached << std::endl;
 
   EXPECT_GT(r4.usage.cachedTokens, 0)
@@ -562,23 +264,18 @@ TEST_F(PrefixCacheE2ETest, CacheReplayScenario) {
   EXPECT_LE(std::abs(r4.usage.cachedTokens - r4ExpectedCached), 1)
       << "Request 4 cached should match block-aligned R4 prompt";
 
-  // -------------------------------------------------------------------------
-  // Request 5: Replay exact R2 prompt → should hit full cache
-  // -------------------------------------------------------------------------
+  // Request 5: Replay exact R2 prompt -> should hit full cache
   std::this_thread::sleep_for(std::chrono::milliseconds(300));
 
   std::cout << "  Request 5 (replay R2 prompt)..." << std::endl;
-  ChatResponse r5 = sendChat(cfg, r2MessagesCopy);
+  ChatResponse r5 = sendChat(r2MessagesCopy);
   ASSERT_TRUE(r5.ok()) << "Request 5 failed: " << r5.error;
   std::cout << "    prompt=" << r5.usage.promptTokens
             << " cached=" << r5.usage.cachedTokens
             << " completion=" << r5.usage.completionTokens << std::endl;
 
-  // R5 replays R2 exactly. R5's tokens match R1's session (prompt + completion)
-  // up to that point, then diverge at the user2 message. The match is limited
-  // to R1's session tokens.
   int r5ExpectedCached = computeExpectedCachedTokens(
-      r1SessionTokens, cfg.firstBlockSize, cfg.blockSize);
+      r1SessionTokens, cfg_.firstBlockSize, cfg_.blockSize);
   std::cout << "    Expected cached: " << r5ExpectedCached << std::endl;
 
   EXPECT_GT(r5.usage.cachedTokens, 0)
@@ -593,21 +290,11 @@ TEST_F(PrefixCacheE2ETest, MultiTurnHashCreation) {
   // Test that multi-turn conversations create proper prefix hashes.
   // Simulates guideLLM multi-turn scenario: each turn builds on the previous,
   // and prefix cache should hit on the shared history.
-  //
-  // Turn 1: system + user1 → cached = 0
-  // Turn 2: system + user1 + assistant1 + user2 → cached > 0 (matches turn 1
-  // prefix) Turn 3: system + user1 + assistant1 + user2 + assistant2 + user3 →
-  // cached > turn2
 
   std::cout << "\n=== Test: Multi-turn hash creation ===" << std::endl;
 
-  auto now = std::chrono::system_clock::now();
-  auto epoch = now.time_since_epoch();
-  auto millis =
-      std::chrono::duration_cast<std::chrono::milliseconds>(epoch).count();
-  // Prepend unique prefix to change the first block hash (prefix cache hashes
-  // from the start, so appending at the end doesn't prevent cache hits)
-  std::string uniquePrefix = "[MULTITURN-TEST-" + std::to_string(millis) + "] ";
+  std::string uniquePrefix =
+      "[" + generateUniqueTestId("MULTITURN-TEST") + "] ";
 
   // Turn 1
   std::vector<Json::Value> messages = {
@@ -615,7 +302,7 @@ TEST_F(PrefixCacheE2ETest, MultiTurnHashCreation) {
       makeMessage("user", "What is a hash table?")};
 
   std::cout << "  Turn 1..." << std::endl;
-  ChatResponse t1 = sendChat(cfg, messages);
+  ChatResponse t1 = sendChat(messages);
   ASSERT_TRUE(t1.ok()) << "Turn 1 failed: " << t1.error;
   std::cout << "    prompt=" << t1.usage.promptTokens
             << " cached=" << t1.usage.cachedTokens << std::endl;
@@ -624,6 +311,7 @@ TEST_F(PrefixCacheE2ETest, MultiTurnHashCreation) {
   int t1Prompt = t1.usage.promptTokens;
 
   std::cout << "t1.content: " << t1.content << std::endl;
+
   // Turn 2
   messages.push_back(makeMessage("assistant", t1.content));
   messages.push_back(makeMessage("user", "How does it handle collisions?"));
@@ -631,17 +319,14 @@ TEST_F(PrefixCacheE2ETest, MultiTurnHashCreation) {
   std::this_thread::sleep_for(std::chrono::milliseconds(300));
 
   std::cout << "  Turn 2..." << std::endl;
-  ChatResponse t2 = sendChat(cfg, messages);
+  ChatResponse t2 = sendChat(messages);
   ASSERT_TRUE(t2.ok()) << "Turn 2 failed: " << t2.error;
   std::cout << "    prompt=" << t2.usage.promptTokens
             << " cached=" << t2.usage.cachedTokens << std::endl;
 
-  // Turn 2 should hit cache on turn 1's session state, which includes both
-  // the original prompt AND completion tokens (including thinking tokens).
-  // The cached amount is block-aligned(prompt + completion) from Turn 1.
   int t1SessionTokens = t1Prompt + t1.usage.completionTokens;
   int t2ExpectedCached = computeExpectedCachedTokens(
-      t1SessionTokens, cfg.firstBlockSize, cfg.blockSize);
+      t1SessionTokens, cfg_.firstBlockSize, cfg_.blockSize);
   std::cout << "    t1 session tokens: " << t1SessionTokens
             << " (prompt=" << t1Prompt
             << " + completion=" << t1.usage.completionTokens << ")"
@@ -652,6 +337,7 @@ TEST_F(PrefixCacheE2ETest, MultiTurnHashCreation) {
       << "Turn 2 cached should match block-aligned turn 1 session";
 
   std::cout << "t2.content: " << t2.content << std::endl;
+
   // Turn 3
   messages.push_back(makeMessage("assistant", t2.content));
   messages.push_back(makeMessage("user", "What about open addressing?"));
@@ -659,15 +345,14 @@ TEST_F(PrefixCacheE2ETest, MultiTurnHashCreation) {
   std::this_thread::sleep_for(std::chrono::milliseconds(300));
 
   std::cout << "  Turn 3..." << std::endl;
-  ChatResponse t3 = sendChat(cfg, messages);
+  ChatResponse t3 = sendChat(messages);
   ASSERT_TRUE(t3.ok()) << "Turn 3 failed: " << t3.error;
   std::cout << "    prompt=" << t3.usage.promptTokens
             << " cached=" << t3.usage.cachedTokens << std::endl;
 
-  // Turn 3 should hit cache on turn 2's session state (prompt + completion).
   int t2SessionTokens = t2.usage.promptTokens + t2.usage.completionTokens;
   int t3ExpectedCached = computeExpectedCachedTokens(
-      t2SessionTokens, cfg.firstBlockSize, cfg.blockSize);
+      t2SessionTokens, cfg_.firstBlockSize, cfg_.blockSize);
   std::cout << "    t2 session tokens: " << t2SessionTokens
             << " (prompt=" << t2.usage.promptTokens
             << " + completion=" << t2.usage.completionTokens << ")"
@@ -684,21 +369,10 @@ TEST_F(PrefixCacheE2ETest, SessionEvictionUnderLoad) {
   // Test that eviction policy works when many sessions are created.
   // Create multiple unique conversations to fill up session slots.
   // Verify that prefix cache still works after eviction occurs.
-  //
-  // Note: This test doesn't directly observe eviction (no API for that),
-  // but it verifies the system remains functional under load and
-  // prefix caching continues to work after many sessions are created.
 
   std::cout << "\n=== Test: Session eviction under load ===" << std::endl;
 
-  auto now = std::chrono::system_clock::now();
-  auto epoch = now.time_since_epoch();
-  auto millis =
-      std::chrono::duration_cast<std::chrono::milliseconds>(epoch).count();
-
-  // Create many unique conversations to trigger potential eviction
-  // Each conversation has a unique system prompt prefix to change
-  // the first block hash and prevent cache hits between conversations
+  int64_t millis = currentTimeMillis();
   constexpr int kNumConversations = 20;
 
   std::cout << "  Creating " << kNumConversations << " unique conversations..."
@@ -712,10 +386,9 @@ TEST_F(PrefixCacheE2ETest, SessionEvictionUnderLoad) {
         makeMessage("system", uniquePrefix + std::string(kSystemPromptCoding)),
         makeMessage("user", "Question " + std::to_string(i))};
 
-    ChatResponse r = sendChat(cfg, messages);
+    ChatResponse r = sendChat(messages);
     ASSERT_TRUE(r.ok()) << "Conversation " << i << " failed: " << r.error;
 
-    // First request for each unique conversation should have cached=0
     EXPECT_EQ(r.usage.cachedTokens, 0)
         << "Conversation " << i << " should have cached=0 (unique prompt)";
   }
@@ -732,7 +405,7 @@ TEST_F(PrefixCacheE2ETest, SessionEvictionUnderLoad) {
       makeMessage("user", "Tell me about whales.")};
 
   std::cout << "  Final conversation - request 1 (fresh)..." << std::endl;
-  ChatResponse f1 = sendChat(cfg, finalMessages);
+  ChatResponse f1 = sendChat(finalMessages);
   ASSERT_TRUE(f1.ok()) << "Final request 1 failed: " << f1.error;
   std::cout << "    prompt=" << f1.usage.promptTokens
             << " cached=" << f1.usage.cachedTokens << std::endl;
@@ -743,13 +416,13 @@ TEST_F(PrefixCacheE2ETest, SessionEvictionUnderLoad) {
   std::this_thread::sleep_for(std::chrono::milliseconds(300));
 
   std::cout << "  Final conversation - request 2 (replay)..." << std::endl;
-  ChatResponse f2 = sendChat(cfg, finalMessages);
+  ChatResponse f2 = sendChat(finalMessages);
   ASSERT_TRUE(f2.ok()) << "Final request 2 failed: " << f2.error;
   std::cout << "    prompt=" << f2.usage.promptTokens
             << " cached=" << f2.usage.cachedTokens << std::endl;
 
   int f2ExpectedCached = computeExpectedCachedTokens(
-      f2.usage.promptTokens, cfg.firstBlockSize, cfg.blockSize);
+      f2.usage.promptTokens, cfg_.firstBlockSize, cfg_.blockSize);
   EXPECT_GT(f2.usage.cachedTokens, 0)
       << "Final R2 should hit cache (system still functional after load)";
   EXPECT_LE(std::abs(f2.usage.cachedTokens - f2ExpectedCached), 1)
